requirements.txt
django~=3.0.0
pillow
pilkit
bootstrap4
django_extensions
django-imagekit
